fix(site/AgentsPage): clean up hook API, fix deprecated types, deduplicate story helper
