+++ site/src/pages/AgentsPage/components/DiffViewer/DiffViewer.stories.tsx
@@ -5,6 +5,7 @@ import { expect, fn, waitFor } from "storybook/test";
 import type { DiffStyle } from "../DiffViewer/DiffViewer";
 import { DiffViewer } from "../DiffViewer/DiffViewer";
 import { InlinePromptInput } from "../DiffViewer/RemoteDiffPanel";
+import { generateLargeDiff } from "./testHelpers";
 
 // biome-ignore format: raw diff string must preserve exact whitespace
 const sampleDiff = [
@@ -433,38 +434,6 @@ export const RenameWithLongPaths: Story = {
 // -------------------------------------------------------------------
 // Large diff for performance testing
 // -------------------------------------------------------------------
-
-/**
- * Generate a realistic unified diff string with many files and lines.
- * Used for benchmarking rendering performance with large diffs.
- */
-function generateLargeDiff(fileCount: number, linesPerFile: number): string {
-	const dirs = ["src", "lib", "utils", "components", "hooks"];
-	const files: string[] = [];
-	for (let f = 0; f < fileCount; f++) {
-		const dir = dirs[f % dirs.length];
-		const fileName = `${dir}/module${f}.ts`;
-		const lines = [
-			`diff --git a/${fileName} b/${fileName}`,
-			`index ${f.toString(16).padStart(7, "0")}..${(f + 1).toString(16).padStart(7, "0")} 100644`,
-			`--- a/${fileName}`,
-			`+++ b/${fileName}`,
-			`@@ -1,${linesPerFile} +1,${linesPerFile + Math.floor(linesPerFile / 10)} @@`,
-		];
-		for (let i = 1; i <= linesPerFile; i++) {
-			lines.push(` // context line ${i} in ${fileName}`);
-			if (i % 10 === 0) {
-				lines.push(`-  const old${i} = getValue(${i});`);
-				lines.push(`+  const new${i} = getUpdatedValue(${i});`);
-			}
-			if (i % 25 === 0) {
-				lines.push(`+  // Added: validation for ${fileName} at line ${i}`);
-			}
-		}
-		files.push(lines.join("\n"));
-	}
-	return files.join("\n");
-}
 
 export const LargeDiff: Story = {
 	args: {
